Refactor: 이미지 서비스 구조 개선
