+++ src/main/java/com/programmers/heycake/domain/comment/facade/CommentFacade.java
@@ -14,7 +14,6 @@ import com.programmers.heycake.domain.comment.service.CommentService;
 import com.programmers.heycake.domain.image.model.dto.ImageResponse;
 import com.programmers.heycake.domain.image.model.dto.ImageResponses;
 import com.programmers.heycake.domain.image.model.vo.ImageType;
-import com.programmers.heycake.domain.image.service.ImageIntegrationService;
 import com.programmers.heycake.domain.image.service.ImageService;
 import com.programmers.heycake.domain.member.model.dto.response.MemberResponse;
 import com.programmers.heycake.domain.member.service.MemberService;
@@ -29,14 +28,13 @@ public class CommentFacade {
 
 	private final MemberService memberService;
 	private final CommentService commentService;
-	private final ImageIntegrationService imageIntegrationService;
 	private final ImageService imageService;
 
 	@Transactional
 	public void deleteComment(Long commentId) {
 		List<ImageResponse> commentImageResponse = imageService.getImages(commentId, ImageType.COMMENT).images();
 		if (!commentImageResponse.isEmpty()) {
-			imageIntegrationService.deleteImages(commentId, ImageType.COMMENT, COMMENT_SUB_PATH);
+			imageService.deleteImages(commentId, ImageType.COMMENT, COMMENT_SUB_PATH);
 		}
 
 		commentService.deleteComment(commentId);
@@ -46,7 +44,7 @@ public class CommentFacade {
 	public void deleteCommentWithoutAuth(Long commentId) {
 		List<ImageResponse> commentImageResponse = imageService.getImages(commentId, ImageType.COMMENT).images();
 		if (!commentImageResponse.isEmpty()) {
-			imageIntegrationService.deleteImages(commentId, ImageType.COMMENT, COMMENT_SUB_PATH);
+			imageService.deleteImages(commentId, ImageType.COMMENT, COMMENT_SUB_PATH);
 		}
 
 		commentService.deleteCommentWithoutAuth(commentId);
@@ -60,7 +58,7 @@ public class CommentFacade {
 				memberId);
 
 		if (commentSaveRequest.image() != null) {
-			imageIntegrationService.createAndUploadImage(commentSaveRequest.image(), COMMENT_SUB_PATH, savedCommentId,
+			imageService.createAndUploadImage(commentSaveRequest.image(), COMMENT_SUB_PATH, savedCommentId,
 					ImageType.COMMENT);
 		}
