+++ src/main/java/com/programmers/heycake/domain/market/facade/EnrollmentFacade.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.programmers.heycake.common.mapper.EnrollmentMapper;
 import com.programmers.heycake.domain.image.model.dto.ImageResponses;
-import com.programmers.heycake.domain.image.service.ImageIntegrationService;
+import com.programmers.heycake.domain.image.service.ImageService;
 import com.programmers.heycake.domain.market.model.dto.request.EnrollmentCreateRequest;
 import com.programmers.heycake.domain.market.model.dto.request.EnrollmentGetListRequest;
 import com.programmers.heycake.domain.market.model.dto.response.EnrollmentDetailNoImageResponse;
@@ -30,7 +30,7 @@ public class EnrollmentFacade {
 	private static final String ENROLLMENT_IMAGE_PATH = "images/marketEnrollments";
 
 	private final EnrollmentService enrollmentService;
-	private final ImageIntegrationService imageIntegrationService;
+	private final ImageService imageService;
 	private final MarketService marketService;
 
 	@Transactional
@@ -38,13 +38,13 @@ public class EnrollmentFacade {
 
 		Long enrollmentId = enrollmentService.createEnrollment(request);
 
-		imageIntegrationService.createAndUploadImage(
+		imageService.createAndUploadImage(
 				request.businessLicenseImage(),
 				ENROLLMENT_IMAGE_PATH,
 				enrollmentId,
 				ENROLLMENT_LICENSE
 		);
-		imageIntegrationService.createAndUploadImage(
+		imageService.createAndUploadImage(
 				request.marketImage(),
 				ENROLLMENT_IMAGE_PATH,
 				enrollmentId,
@@ -57,7 +57,7 @@ public class EnrollmentFacade {
 	@Transactional(readOnly = true)
 	public EnrollmentDetailWithImageResponse getMarketEnrollment(Long enrollmentId) {
 		EnrollmentDetailNoImageResponse enrollment = enrollmentService.getMarketEnrollment(enrollmentId);
-		ImageResponses images = imageIntegrationService.getImages(enrollmentId, ENROLLMENT_MARKET);
+		ImageResponses images = imageService.getImages(enrollmentId, ENROLLMENT_MARKET);
 		return EnrollmentMapper.toEnrollmentDetailWithImageResponse(enrollment, images);
 	}
 
@@ -75,7 +75,7 @@ public class EnrollmentFacade {
 		List<EnrollmentListSummaryNoImageResponse> noImageResponses = enrollmentService.getMarketEnrollments(request);
 		List<EnrollmentListSummaryWithImageResponse> withImageResponses = noImageResponses.stream()
 				.map(enrollment -> {
-					ImageResponses images = imageIntegrationService.getImages(enrollment.enrollmentId(), ENROLLMENT_MARKET);
+					ImageResponses images = imageService.getImages(enrollment.enrollmentId(), ENROLLMENT_MARKET);
 					return EnrollmentMapper.toEnrollmentListSummaryWithImageResponse(enrollment, images);
 				})
 				.toList();
