+++ src/main/java/com/programmers/heycake/domain/offer/facade/OfferFacade.java
@@ -11,7 +11,6 @@ import org.springframework.transaction.annotation.Transactional;
 import com.programmers.heycake.common.mapper.OfferMapper;
 import com.programmers.heycake.domain.comment.facade.CommentFacade;
 import com.programmers.heycake.domain.image.model.dto.ImageResponses;
-import com.programmers.heycake.domain.image.service.ImageIntegrationService;
 import com.programmers.heycake.domain.image.service.ImageService;
 import com.programmers.heycake.domain.market.model.dto.response.MarketDetailNoImageResponse;
 import com.programmers.heycake.domain.market.service.MarketService;
@@ -33,10 +32,9 @@ public class OfferFacade {
 	private final OfferService offerService;
 	private final MarketService marketService;
 	private final MemberService memberService;
-	private final ImageService imageService;
 	private final CommentFacade commentFacade;
 	private final HistoryService historyService;
-	private final ImageIntegrationService imageIntegrationService;
+	private final ImageService imageService;
 
 	@Transactional
 	public Long saveOffer(OfferSaveRequest offerSaveRequest) {
@@ -44,7 +42,7 @@ public class OfferFacade {
 		Long savedOfferId = offerService.saveOffer(offerSaveRequest.orderId(), offerSaveRequest.expectedPrice(),
 				offerSaveRequest.content());
 
-		imageIntegrationService.createAndUploadImage(offerSaveRequest.offerImage(), OFFER_IMAGE_SUB_PATH, savedOfferId,
+		imageService.createAndUploadImage(offerSaveRequest.offerImage(), OFFER_IMAGE_SUB_PATH, savedOfferId,
 				OFFER);
 
 		return savedOfferId;
@@ -53,7 +51,7 @@ public class OfferFacade {
 	@Transactional
 	public void deleteOffer(Long offerId) {
 		Long marketId = marketService.getMarketIdByMember(memberService.getMemberById(getMemberId()));
-		imageIntegrationService.deleteImages(offerId, OFFER, OFFER_IMAGE_SUB_PATH);
+		imageService.deleteImages(offerId, OFFER, OFFER_IMAGE_SUB_PATH);
 
 		List<Long> offerCommentIdList = offerService.getOfferCommentIdList(offerId);
 		offerCommentIdList.forEach(commentFacade::deleteCommentWithoutAuth);
@@ -63,7 +61,7 @@ public class OfferFacade {
 
 	@Transactional
 	public void deleteOfferWithoutAuth(Long offerId) {
-		imageIntegrationService.deleteImages(offerId, OFFER, OFFER_IMAGE_SUB_PATH);
+		imageService.deleteImages(offerId, OFFER, OFFER_IMAGE_SUB_PATH);
 
 		List<Long> offerCommentIdList = offerService.getOfferCommentIdList(offerId);
 		offerCommentIdList.forEach(commentFacade::deleteCommentWithoutAuth);
